split DocumentEditor/link into shared/ui
