+++ packages/fields-document/src/DocumentEditor/index.tsx
@@ -24,7 +24,7 @@ import { type EditableProps } from 'slate-react/dist/components/editable'
 import { type ComponentBlock } from '../component-blocks'
 import { type DocumentFeatures } from '../views'
 import { withParagraphs } from './paragraphs'
-import { withLink, wrapLink } from './link'
+import { withLink, wrapLink } from './link-shared'
 import { withLayouts } from './layouts'
 import { clearFormatting, type Mark } from './utils'
 import { Toolbar } from './Toolbar'
